+++ src/control_plane/configuration/mod.rs
@@ -1,0 +1,12 @@
+mod config_parse_error;
+mod parse_project_config;
+mod raw_project_config;
+mod raw_service_config;
+
+pub(crate) use config_parse_error::ConfigParseError;
+pub(crate) use parse_project_config::parse_project_config;
+pub(crate) use raw_project_config::RawProjectConfig;
+pub(crate) use raw_service_config::RawServiceConfig;
+
+#[cfg(test)]
+mod tests;